+++ packages/multichain-account-service/src/index.ts
@@ -25,7 +25,6 @@ export type {
 export {
   AccountProviderWrapper,
   BaseBip44AccountProvider,
-  DeleteAccountsError,
   SnapAccountProvider,
   TimeoutError,
   EVM_ACCOUNT_PROVIDER_NAME,
@@ -39,7 +38,10 @@ export {
   XLM_ACCOUNT_PROVIDER_NAME,
   XlmAccountProvider,
 } from './providers/index.js';
-export type { DeleteAccountsFailure } from './providers/index.js';
+export type {
+  DeleteAccountsFailure,
+  DeleteAccountsResult,
+} from './providers/index.js';
 export { MultichainAccountWallet } from './MultichainAccountWallet.js';
 export { MultichainAccountGroup } from './MultichainAccountGroup.js';
 export { MultichainAccountService } from './MultichainAccountService.js';